requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
dependency-injector==4.48.3
fastapi==0.128.0
greenlet==3.3.0
h11==0.16.0
httpx==0.28.1
idna==3.11
psycopg==3.2.3
psycopg-binary==3.2.3
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pytest==8.3.4
pytest-asyncio==0.25.0
python-dotenv==1.2.1
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ulid-py==1.1.0
uvicorn==0.40.0
